Feature/#183: FloatingTitleNumberField 수정
