--- Harubee/Sources/Presentation/Common/Components/FloatingTitleTextField.swift
@@ -1,160 +1,0 @@
-//
-//  FloatingTitleTextField.swift
-//  Harubee-iOS
-//
-//  Created by Seo-Jooyoung on 10/30/24.
-//  Copyright © 2024 namudiEtc. All rights reserved.
-//
-
-import SwiftUI
-
-struct FloatingTitleTextField: View {
-  var title: String
-  var shouldShowKeyboard: Bool = false
-  
-  @Binding var text: String
-  @FocusState private var isTextfieldFocused: Bool
-  
-  var body: some View {
-    VStack(spacing: 0) {
-      Text(title)
-        .font(.pretendardMedium_12)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .foregroundStyle(
-          !text.isEmpty
-          ? (isTextfieldFocused ? .main : .textBright)
-          : .clear
-        )
-        .offset(y: !text.isEmpty ? -4 : 0)
-        .animation(.easeOut(duration: 0.2), value: !text.isEmpty)
-        .padding(.leading, 4)
-      
-      TextField(title, text: $text)
-        .frame(maxWidth: .infinity)
-        .font(.pretendardMedium_18)
-        .focused($isTextfieldFocused)
-        .padding(.leading, 4)
-      
-      Rectangle()
-        .frame(height: !isTextfieldFocused ? 1 : 2)
-        .foregroundStyle(
-          !isTextfieldFocused
-          ? .textBrighter
-          : .mainBright
-        )
-        .padding(.top, !isTextfieldFocused ? 8 : 7)
-    }
-    .onAppear {
-        self.isTextfieldFocused = shouldShowKeyboard
-    }
-    .contentShape(Rectangle())
-  }
-}
-
-struct FloatingTitleNumberField: View {
-  
-  enum TextSize {
-    case large
-    case medium
-  }
-  
-  let title: String
-  let textSize: TextSize
-  @Binding var text: String
-  @Binding var isFocused: Bool
-  var isErrorTextVisible: Bool? = nil
-  
-  private var textFont: Font {
-    if textSize == .large {
-      .pretendardSemibold_28
-    } else {
-      .pretendardMedium_18
-    }
-  }
-  
-  var body: some View {
-    VStack(alignment: .leading, spacing: 0) {
-      Text(title)
-        .font(.pretendardMedium_12)
-        .frame(maxWidth: .infinity, alignment: .leading)
-        .foregroundStyle(
-          text.isEmpty
-          ? .clear
-          : (
-            isFocused
-            ? (isErrorTextVisible ?? false ? .redDefault : .main)
-            : .textBright
-          )
-        )
-        .offset(y: !text.isEmpty ? -2 : 0)
-        .animation(.easeOut(duration: 0.2), value: !text.isEmpty)
-        .padding(.leading, 4)
-      
-      Text(text.isEmpty ? title : text)
-        .foregroundStyle(
-          !text.isEmpty ? .textBlack : .placeholder
-        )
-        .font(textFont)
-        .padding(.leading, 4)
-      
-      Rectangle()
-        .frame(height: isFocused ? 2 : 1)
-        .foregroundStyle(
-          isFocused
-          ? (isErrorTextVisible ?? false && !text.isEmpty
-             ? .redDefault
-             : .mainBright)
-          : .textBrighter
-        )
-        .padding(.top, isFocused ? 7 : 8)
-      
-      if isErrorTextVisible ?? false {
-        errorText
-          .foregroundStyle(
-            text.isEmpty
-            ? .clear
-            : .redDefault
-          )
-          .offset(y: !text.isEmpty ? 4 : 0)
-          .animation(.easeOut(duration: 0.2), value: !text.isEmpty)
-          .padding(.leading, 4)
-      }
-    }
-    .frame(maxWidth: .infinity)
-    .contentShape(Rectangle())
-  }
-  
-  private var errorText: some View {
-    HStack(spacing: 2) {
-      Image(systemName: "exclamationmark.circle")
-        .font(.system(size: 12, weight: .regular))
-      Text("0보다 큰 금액을 입력해주세요")
-        .font(.pretendardSemibold_12)
-    }
-  }
-}
-
-#Preview {
-  @Previewable @State var text = ""
-  @Previewable @State var isFocused = false
-  
-  VStack {
-    Spacer()
-    FloatingTitleTextField(title: "Title", text: $text)
-    FloatingTitleNumberField(
-      title: "Test",
-      textSize: .medium,
-      text: $text,
-      isFocused: $isFocused
-    )
-      .onTapGesture {
-        isFocused = true
-      }
-    Spacer()
-  }
-  .contentShape(Rectangle())
-  .onTapGesture {
-    isFocused = false
-  }
-  
-}
